Find how many times letter occurs in word.py
You can find how many times a particular letter occurs in the word. For example, if we enter apple and enter the letter as p, it will say that there are 2 p's in the word apple. But if we give the input as apple and enter the letter as d, it will say that there is no d in the word apple.